+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to http://semver.org/spec/v2.0.0.html.
 
 ## [unreleased]
 
+- Add tt.LogLevel with info and debug
 - Remove tt.ackman
 
 ## [0.15.0] 2022-10-01